Fix bug with ordering of users
Because we fetch users information on-the-fly, the ordering of the users
in the users array was just not sorted at all. We sort the users after
we receive a new user in the app.

Also, binary search now fallsback to a linear search in case of failure,
just in case there is another bug of this sort.